Implemented collisions method

+++ src/math/collisionMath.cpp
@@ -44,7 +44,7 @@
 // 5/15/2023
 // (the coefficient of) restitution should be between 0 and 1, 
 // 0 is inelastic, 1 is elastic
-void collide(ObjectNd* obj1, ObjectNd* obj2, float restitution){
+void collide(ObjectNd* obj1, ObjectNd* obj2, double restitution){
     // 5/24/2023
     // the initial masses
     // pointers are really nice
